Reorganized Customer block test cases + new tests for DTO factories
