+++ tests/Customer/CustomerDataFactoryTest.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Grixu\SociusModels\Tests\Customer;
+
+use Grixu\SociusModels\Customer\Factories\CustomerDataFactory;
+use Grixu\SociusModels\Tests\Helpers\DataFactoryTestCase;
+
+class CustomerDataFactoryTest extends DataFactoryTestCase
+{
+    protected string $factory = CustomerDataFactory::class;
+}
