+++ tests/Customer/CustomerMigrationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Grixu\SociusModels\Tests;
+namespace Grixu\SociusModels\Tests\Customer;
 
 use CreateCustomersTable;
 use Grixu\SociusModels\Tests\Helpers\MigrationTest;
@@ -13,7 +13,7 @@ class CustomerMigrationTest extends MigrationTest
     {
         parent::setUp();
 
-        require_once __DIR__ . '/../migrations/customer/2020_09_30_102037_create_customers_table.php';
+        require_once __DIR__ . '/../../migrations/customer/2020_09_30_102037_create_customers_table.php';
         $this->obj = new CreateCustomersTable();
     }
 }
